move holiday calculations to seperate file
